// src/app/models/transaction-history.interface.ts
export interface TransactionHistory {
  TransactionNo: string;
  BeneficiaryID: string;
  BeneficiaryName: string;
  BankName: string;
  BankBranch: string;
  AccNo: string;
  Date: string;
  Currency: string;
  AmountFC: string;
  AmountKD: string;
  TotalKD: string;
  BenefColor: string;
  Status: string;
}